Small wording changes
Added a notice about italics.  Not sure if you want to spell out IDE or make it a blue word.

I moved the statements about pre/post being on the exam and software validity not to the beginning of the part on validity so it's easier for teachers to skip.